CDA/WEB_FLASK/app/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
configobj==5.0.6
cycler==0.10.0
dbus-python==1.2.16
distro-info==1.0+deb11u1
flask==3.0.3
flask-wtf==1.2.1
fuse-python==1.0.2
gpg===1.14.0-unknown
httplib2==0.18.1
idna==2.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
kiwisolver==1.3.1
MarkupSafe==2.1.5
matplotlib==3.3.4
mysql-connector==2.2.9
mysql-connector-python==9.0.0
numpy==1.19.5
olefile==0.46
Pillow==8.1.2
Pivy==0.6.5
ply==3.11
psutil==5.8.0
pycairo==1.16.2
pycurl==7.43.0.6
PyGObject==3.38.0
pylibacl==0.6.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
PySimpleSOAP==1.16.2
python-apt==2.2.1
python-dateutil==2.8.1
python-debian==0.1.39
python-debianbts==3.1.0
python-dotenv==1.0.1
pyxattr==0.7.2
PyYAML==5.3.1
reportbug==7.10.3+deb11u2
requests==2.25.1
six==1.16.0
terminator==2.1.0
tornado==6.1
unattended-upgrades==0.1
urllib3==1.26.5
werkzeug==3.0.4
wtforms==3.1.2
zipp==3.20.2
